Make TrieLogManager the default implementation and remove the interface Signed-off-by: Simon Dudley <simon.dudley@consensys.net>
